+++ src/check/scanners/mod.rs
@@ -7,6 +7,7 @@ pub(crate) mod empty_section_title;
 pub(crate) mod footnotes;
 pub(crate) mod illegal_sections;
 pub(crate) mod links;
+pub(crate) mod obsolete_occurrences;
 pub(crate) mod section_capitalization;
 pub(crate) mod section_level;
 pub(crate) mod unordered_sections;
